+++ visualization/app/codeCharta/state/effects/resetChosenMetrics/resetChosenMetrics.effect.ts
@@ -1,11 +1,16 @@
 import { Inject, Injectable } from "@angular/core"
-import { tap } from "rxjs"
-import { isAnyMetricAvailable } from "./utils/metricHelper"
+import { filter, tap } from "rxjs"
+import { defaultNMetrics, isAnyMetricAvailable, areScenarioSettingsApplicable } from "./utils/metricHelper"
 import { createEffect } from "../../angular-redux/effects/createEffect"
 import { Store } from "../../angular-redux/store"
 import { nodeMetricDataSelector } from "../../selectors/accumulatedData/metricData/nodeMetricData.selector"
+import { setState } from "../../store/state.actions"
+import { ScenarioHelper } from "../../../util/scenarioHelper"
 import { setDistributionMetric } from "../../store/dynamicSettings/distributionMetric/distributionMetric.actions"
 import { getDefaultDistribution } from "./utils/getDefaultDistributionMetric"
+import { setAreaMetric } from "../../store/dynamicSettings/areaMetric/areaMetric.actions"
+import { setHeightMetric } from "../../store/dynamicSettings/heightMetric/heightMetric.actions"
+import { setColorMetric } from "../../store/dynamicSettings/colorMetric/colorMetric.actions"
 
 @Injectable()
 export class ResetChosenMetricsEffect {
@@ -14,10 +19,18 @@ export class ResetChosenMetricsEffect {
 	resetChosenDistributionMetric$ = createEffect(
 		() =>
 			this.store.select(nodeMetricDataSelector).pipe(
+				filter(isAnyMetricAvailable),
 				tap(nodeMetricData => {
-					if (isAnyMetricAvailable(nodeMetricData)) {
-						// when migrating area, height and color service, their resetting will be added here as well
-						this.store.dispatch(setDistributionMetric(getDefaultDistribution(nodeMetricData)))
+					this.store.dispatch(setDistributionMetric(getDefaultDistribution(nodeMetricData)))
+
+					const defaultScenarioSettings = ScenarioHelper.getDefaultScenarioSetting()
+					if (areScenarioSettingsApplicable(defaultScenarioSettings, nodeMetricData)) {
+						this.store.dispatch(setState(defaultScenarioSettings))
+					} else {
+						const [defaultedAreaMetric, defaultedHeightMetric, defaultedColorMetric] = defaultNMetrics(nodeMetricData, 3)
+						this.store.dispatch(setAreaMetric(defaultedAreaMetric))
+						this.store.dispatch(setHeightMetric(defaultedHeightMetric))
+						this.store.dispatch(setColorMetric(defaultedColorMetric))
 					}
 				})
 			),